requirements.txt
streamlit>=1.32.0
chromadb>=0.4.24
pydantic>=2.6.0
fastapi>=0.100.0
uvicorn>=0.23.0
twilio>=8.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
langchain-google-genai>=1.0.3
langgraph>=0.0.39
langchain-core>=0.1.45
pandas>=2.0.0